feat(no-implicit-any-catch)!: default to allow explicit any
BREAKING CHANGE: allowExplicitAny now defaults to `true` instead of `false.

This change was made to align with other rules in this codebase dealing with explicit `any` (namely `throw-error`).  Additionally, lax codebases commonly turn on this setting, or else turn off this rule entirely.  By relaxing the default options, developers will be encouraged to use this rule, and stricter codebases can re-opt-in to banning `any`.  The upcoming `strict` config will set this option back to `false` accordingly.