feat(no-misused-observables): new rule (#58)
New rule based on `@typescript-eslint/no-misused-promises`. Ensures you
don't return an Observable where void is expected, or try to spread an
Observable. This rule is added to the `strict` configuration.

Resolves #43 .

Note: `checksVoidReturn` is noticeably slow, specifically stemming from
the `arguments`, `properties`, and `variables` sub-options. It's still
faster than `@typescript-eslint/no-deprecated` and some `import-x`
rules, but it's in the top 10 with `TIMING=1` enabled. But since the
implementation is similar to `@typescript-eslint/no-misused-promises`
which has similar performance, we'll allow it.