feat(no-topromise): suggest an exact replacement for toPromise() (#184)
This adds the "most correct" suggestion to `no-topromise` for replacing
`toPromise()`.

The old `toPromise()` would resolve to `undefined`, while the new
`lastValueFrom` rejects with `EmptyError`. So if a user followed the old
top suggestion and replaced `toPromise` with `lastValueFrom`, they might
be surprised to find `EmptyError` now thrown instead. This wouldn't be a
problem for someone using the `strict` config, which includes
`no-ignored-default-value`, but users of the `recommended` config
wouldn't get that lint notification.

Resolves #168