feat(no-unnecessary-collection): new rule (#281)
New rule to detect unnecessary usage of `combineLatest`, `forkJoin`,
etc. where a collection (object or array) containing only a single
observable is passed in. In these cases, `combineLatest`, etc. can be
replaced with the single observable itself (after accounting for slight
changes in output).

* This rule is added to the strict config.
* This rule does not include a fixer. A fixer could be implemented in
future for `merge`, `race`, `concat`, and the overload of
`combineLatest` that accepts an array.
* This rule does not support variable-length arrays and non-literal
objects because their lengths are not guaranteed.
* Like most rules in this plugin, this rule does not support import
aliases.

Resolves #39 and cartant#107 .